# perf(translation): improve if condition to return all values except `None`