Improves #16949 

### Before


```python

In [3]: %timeit frappe._("Sales Invoice", "en")
9.33 µs ± 556 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)


```

### After

```python

In [3]: %timeit frappe._("Sales Invoice", "en")
2.79 µs ± 7.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

```