@ankush 2022-05-23
Assuming, this is because the language dict can be empty in case of English and no user translations.