@barredterra 2022-11-23
@sagarvora this lead to a bug in our app [PDF on Submit](https://github.com/alyf-de/erpnext_pdf-on-submit).

https://github.com/frappe/frappe/blob/fd8cfd3058928fad5d32df4ca5e25a1dd50afecb/frappe/www/printview.py#L224

When `_()` gets called from inside `template.render()`, `frappe.local.lang_full_dict` is an empty dict initially. Before your change, this evaluated to false and the correct translations got loaded. after your changes, it returns the empty dict and, thus, doesn't translate.